Tried to implement a 'continuous angular grid', but it either does not work or is too slow compared to the current method. Also added error messages whenever invalid option combinations are chosen for symmeterization, quadrature convergence study, and the type of Hv Calculator

+++ Hv-calculators/linRotor/cartSphBasis/potentialCalc/CommonRoutines/gridFcns.cpp
@@ -48,7 +48,7 @@ universeProp* generateGrid(int numDim, double *gridMax, int *gridPoints) {
 		universe->grid[i].DIM(universe->numDim);
 		for (j=0; j<universe->numDim; j++) {
 			//universe->grid[i].COOR(j) = indexDim(j, i, universe->numDim, universe->grid_num)*universe->d_i[j]; //Seems to work, but not fully tested.
-			universe->grid[i].COOR(j) = (indexDim(j, i, universe->numDim, universe->grid_num)*universe->d_i[j]) - (universe->grid_max[j]/2.0); //Seems to work, but not fully tested.; subtract 1/2*box_i_max to centre origin
+			universe->grid[i].COOR(j) = (double(indexDim(j, i, universe->numDim, universe->grid_num))*universe->d_i[j]) - (universe->grid_max[j]/2.0); //Seems to work, but not fully tested.; subtract 1/2*box_i_max to centre origin
 		}		
 	}
 	}
